# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Water-id Translator Documentation",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "admin@water-id.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "2.0.0"
    },
    "servers": [
        {
            "url": "http://translator.local/api/v1",
            "description": "Water-id Translator API Server"
        }
    ],
    "paths": {
...
        "/translations": {
            "get": {
                "tags": [
                    "Translations"
                ],
                "summary": "Get list of translations",
                "description": "Returns list of translations",
                "operationId": "getTranslationsList",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    },
                    {
                        "name": "perPage",
                        "in": "query",
                        "description": "Number of item per page",
                        "schema": {
                            "type": "integer"
                        },
                        "example": 20
                    },
                    {
                        "name": "sortBy",
                        "in": "query",
                        "description": "Column name to sort by",
                        "schema": {
                            "type": "string"
                        },
                        "example": "id"
                    },
                    {
                        "name": "sortDesc",
                        "in": "query",
                        "description": "Column name to sort descending",
                        "required": true,
                        "schema": {
                            "type": "boolean"
                        },
                        "example": true
                    },
                    {
                        "name": "q",
                        "in": "query",
                        "description": "Search Query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "translation",
                        "in": "query",
                        "description": "Translation",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "phrase",
                        "in": "query",
                        "description": "Phrase",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "phraseId",
                        "in": "query",
                        "description": "Phrase Id",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "language",
                        "in": "query",
                        "description": "Language Title",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "languageId",
                        "in": "query",
                        "description": "Language Id",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "user",
                        "in": "query",
                        "description": "User Name",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "userId",
                        "in": "query",
                        "description": "User Id",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TranslationResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiUnAuthException"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Invalid scope(s) provided.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiAccessDeniedException"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Translations"
                ],
                "summary": "Store new translation",
                "description": "Returns translation data",
                "operationId": "storeTranslation",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreTranslationRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TranslationOut"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiRequestException"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiUnAuthException"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Invalid scope(s) provided.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiAccessDeniedException"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiUnprocessableEntityException"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport": [
                            "*"
                        ]
                    }
                ]
            }
        },
        "/translations/{id}": {
            "get": {
                "tags": [
                    "Translations"
                ],
                "summary": "Get translation information",
                "description": "Returns translation data",
                "operationId": "getTranslationById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Translation id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TranslationOut"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiRequestException"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiUnAuthException"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Invalid scope(s) provided.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiAccessDeniedException"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Resource Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiNotFoundException"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport": [
                            "*"
                        ]
                    }
                ]
            },
            "put": {
                "tags": [
                    "Translations"
                ],
                "summary": "Update existing translation",
                "description": "Returns updated translation data",
                "operationId": "updateTranslation",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Translation id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateTranslationRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TranslationOut"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiRequestException"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiUnAuthException"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Invalid scope(s) provided.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiAccessDeniedException"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Resource Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiNotFoundException"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport": [
                            "*"
                        ]
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Translations"
                ],
                "summary": "Delete existing translation",
                "description": "Deletes a record and returns no content",
                "operationId": "deleteTranslation",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Translation id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiUnAuthException"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Invalid scope(s) provided.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiAccessDeniedException"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Resource Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiNotFoundException"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport": [
                            "*"
                        ]
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "ApiAccessDeniedException": {
                "properties": {
                    "message": {
                        "description": "The err message",
                        "type": "string",
                        "example": "AccessDenied"
                    }
                },
                "type": "object"
            },
...
            "ApiUnAuthException": {
                "properties": {
                    "message": {
                        "description": "The err message",
                        "type": "string",
                        "example": "Unauthenticated"
                    }
                },
                "type": "object"
            },
...
            "StoreTranslationRequest": {
                "title": "Store Translation request",
                "description": "Store Translation request body data",
                "required": [
                    "title",
                    "iso_code",
                    "is_primary",
                    "text_direction",
                    "active"
                ],
                "properties": {
                    "translation": {
                        "title": "Translation",
                        "description": "Translation",
                        "type": "string",
                        "example": "A new translation"
                    },
                    "phrase_id": {
                        "title": "PhraseID",
                        "description": "Phrase ID",
                        "type": "integer",
                        "example": "A phrase id"
                    },
                    "language_id": {
                        "title": "LanguageID",
                        "description": "Language ID",
                        "type": "integer",
                        "example": "A language id"
                    },
                    "user_id": {
                        "title": "UserID",
                        "description": "User ID",
                        "type": "integer",
                        "example": "A user id"
                    }
                },
                "type": "object"
            },
...
            "TranslationResource": {
                "title": "TranslationResource",
                "description": "Translation resource",
                "properties": {
                    "data": {
                        "title": "Data",
                        "description": "Data wrapper",
                        "type": "array",
                        "items": {
                            "ref": null
                        }
                    }
                },
                "type": "object",
                "xml": {
                    "name": "TranslationResource"
                }
            },
...
            "TranslationOut": {
                "title": "TranslationOut",
                "description": "TranslationOut model",
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "ID",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "translation": {
                        "title": "Translation",
                        "description": "Translation",
                        "type": "string",
                        "example": "A new translation"
                    },
                    "phrase_id": {
                        "title": "Phrase ID",
                        "description": "Phrase ID",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "language_id": {
                        "title": "Language ID",
                        "description": "Language ID",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "user_id": {
                        "title": "User ID",
                        "description": "User ID",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "language": {
                        "$ref": "#/components/schemas/LanguageOut"
                    },
                    "phrase": {
                        "$ref": "#/components/schemas/PhraseOut"
                    },
                    "user": {
                        "$ref": "#/components/schemas/UserOut"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "TranslationOut"
                }
            },
...
        },
        "securitySchemes": {
            "passport": {
                "type": "apiKey",
                "description": "Enter \"Bearer + [space] + [token]\" in the text input below(Value).",
                "example": "Bearer 1qaz2wsx3edc",
                "in": "header",
                "name": "Authorization"
            }
        }
    },
...
}